+++ Terrafrom-AWS-Prasanth/terraform-aws-cognito/data.tf
@@ -3,5 +3,7 @@ data "aws_caller_identity" "current" {}
 data "aws_iam_account_alias" "current" {}
 
 data "local_file" "saml_metadata" {
+  count = var.saml_metadata_content == null ? 1 : 0
+
   filename = var.samlmetadatafile
 }