+++ Terrafrom-AWS-Prasanth/terraform-aws-cognito/main.tf
@@ -1,5 +1,9 @@
 locals {
   account_alias = data.aws_iam_account_alias.current.account_alias
+  saml_metadata = coalesce(
+    var.saml_metadata_content,
+    one(data.local_file.saml_metadata[*].content),
+  )
   base_attribute_mapping = {
     "username"      = "SAMLAccountName"
     "custom:groups" = "groups"
@@ -102,7 +106,7 @@ resource "aws_cognito_identity_provider" "main" {
   provider_type = "SAML"
 
   provider_details = {
-    MetadataFile = data.local_file.saml_metadata.content
+    MetadataFile = local.saml_metadata
   }
 
   attribute_mapping = merge(